--- flextest.html
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width,initial-scale=1.0">
-    <title>Flex test</title>
-    <link href="resources/main.css" rel="stylesheet">
-    <script src="scripts/newapp.js" type="text/javascript"></script>
-</head>
-<body>
-    <div class="container loading" id="main">
-        <div class="toolbar">
-            <div class="menu-title">Raycaster</div>
-            <menu data-ref="main-menu"></menu>
-            <div class="spacer"></div>
-            <div class="btn-activecolor-primary small view-selector" data-ref="view-selector">
-                <button value="editor"><span>Editor</span></button>
-                <button value="split"><span>Split</span></button>
-                <button value="view"><span>View</span></button>
-            </div>
-        </div>
-        <div class="window-container">
-            <div class="window window-left editor">
-                <div class="editor-content"><canvas id="panel" tabindex='1'></canvas></div>
-                <div class="editor-footer"><button value="hello world" data-ref="btn-test" style="display: none;"><span>Test</span></button></div>
-            </div>
-            <div class="window window-right">
-                <div class="display-container">
-                    <div class="display-container-inner" id="raycast-view"></div>
-                </div>
-            </div>
-        </div>
-    </div>
-</body>
-</html>
