+++ index.html
@@ -4,84 +4,31 @@
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width,initial-scale=1.0">
-    <title>Raycasting</title>
-    <link href="resources/style.css" rel="stylesheet">
+    <title>Flex test</title>
+    <link href="resources/main.css" rel="stylesheet">
     <script src="scripts/app.js" type="text/javascript"></script>
 </head>
-<body onload="init()">
-    <div class="content">
-        <div class="d-flex" style="background-color: #ddd;">
-            <div class="field pa-2 d-flex align-center">
-                <label for="optionsColor">Color:</label>
-                <select name="optionsColor" id="optionsColor">
-                    <option value="">Random</option>
-                    <option value="#00000000">Transparent</option>
-                    <option value="#444">Grey</option>
-                    <option value="#f00">Red</option>
-                    <option value="#0F0">Green</option>
-                    <option value="#00f">Blue</option>
-                    <option value="#000">Black</option>
-                    <option value="#09f">LightBlue</option>
-                </select>
-            </div>
-            <div class="field pa-2 d-flex align-center">
-                <label for="optionsTexture">Texture:</label>
-                <select name="optionsTexture" id="optionsTexture">
-                    <option value="">None</option>
-                    <option value="concrete1">Concrete 1</option>
-                    <option value="concrete_pillar">Concrete Pillar</option>
-                    <option value="grate">Grate</option>
-                    <option value="debug">Debug</option>
-                </select>
-            </div>
-            <div class="field pa-2 d-flex align-center">
-                <label for="optionsIsMirror">Mirror:</label>
-                <input type="checkbox" name="optionsIsMirror" id="optionsIsMirror" />
-            </div>
-            <div class="field pa-2 d-flex align-center">
-                <label for="optionsOpaque">Opaque:</label>
-                <input type="checkbox" name="optionsOpaque" id="optionsOpaque" checked />
-            </div>
-            <div class="field pa-2 d-flex align-center">
-                <label for="optionsOpacity">Opacity:</label>
-                <input type="range" name="optionsOpacity" id="optionsOpacity" min="0" max="1" value="1" step="0.1" class="slider" id="myRange">
-                <span id="optionsOpacityValue">100%</span>
-            </div>
-            <div class="field pa-2 d-flex align-center">
-                <label for="optionsSolid">Solid:</label>
-                <input type="checkbox" name="optionsSolid" id="optionsSolid" checked />
-            </div>
+<body>
+    <div class="container loading" id="main">
+        <div class="toolbar">
+            <div class="menu-title">Raycaster</div>
+            <menu data-ref="main-menu"></menu>
             <div class="spacer"></div>
-            <div class="field pa-2 d-flex align-center">
-                <button onclick="actionToggleDialog()">ToggleDialog</button>
+            <div class="btn-activecolor-primary small view-selector" data-ref="view-selector">
+                <button value="editor"><span>Editor</span></button>
+                <button value="split"><span>Split</span></button>
+                <button value="view"><span>View</span></button>
             </div>
         </div>
-        <div id="engine-window">
-            <div class="field">
-                <canvas id="panel" tabindex='1'></canvas>
+        <div class="window-container">
+            <div class="window window-left editor">
+                <div class="editor-content"><canvas id="panel" tabindex='1'></canvas></div>
+                <div class="editor-footer"><button value="hello world" data-ref="btn-test" style="display: none;"><span>Test</span></button></div>
             </div>
-            <div class="d-flex gray">
-                <div class="field pa-3 br"><strong class="legend-input">SPACE</strong>Drag map</div>
-                <div class="field pa-3 br"><strong class="legend-input">+ / -</strong>Increase / Decrease Gridsize</div>
-            </div>
-        </div>
-        <div class="hide" id="export-dialog">
-            <div class="d-flex align-center bb">
-                <div class="field pa-2">
-                    <button onclick="actionDownload()">Download</button>
+            <div class="window window-right">
+                <div class="display-container">
+                    <div class="display-container-inner" id="raycast-view"></div>
                 </div>
-                <div class="field pa-2 d-flex align-center">
-                    <label for="filejson">Import</label>
-                    <input type="file" name="filejson" id="filejson" />
-                </div>
-            </div>
-            <div class="field pa-2">
-                <label for="datajson">Json data:</label>
-                <textarea name="datajson" id="datajson"></textarea>
-            </div>
-            <div class="field pa-2 pt-0">
-                <button onclick="actionExport()">Export</button>
-                <button onclick="actionImport()">Import</button>
             </div>
         </div>
     </div>
