--- resources/style.css
@@ -1,154 +1,0 @@
-html, body {
-    margin:0;
-    padding:0;
-    background-color: #eee;
-    font-family: Arial, Helvetica, sans-serif;
-}
-div {
-    border-color: inherit;
-}
-.content {
-    margin: 50px auto;
-    width: 1024px;
-    background-color: #fff;
-    box-shadow: rgba(0,0,0,0.3) 1px 1px 8px;
-    border-color: #ddd;
-}
-
-.gray {
-    background-color: #ddd;
-    border-color: #aaa;
-}
-
-.bb {
-    border-bottom: 1px solid;
-    border-color: inherit;
-}
-.br {
-    border-right: 1px solid;
-    border-color: inherit;
-}
-
-.content .pa-0 {padding: 0;}
-.content .pa-1 {padding: 4px;}
-.content .pa-2 {padding: 8px;}
-.content .pa-3 {padding: 12px;}
-.content .pa-4 {padding: 16px;}
-.content .pa-5 {padding: 20px;}
-
-.content .pt-0 {padding-top: 0;}
-.content .pt-1 {padding-top: 4px;}
-.content .pt-2 {padding-top: 8px;}
-.content .pt-3 {padding-top: 12px;}
-.content .pt-4 {padding-top: 16px;}
-.content .pt-5 {padding-top: 20px;}
-
-.content .pl-0 {padding-left: 0;}
-.content .pl-1 {padding-left: 4px;}
-.content .pl-2 {padding-left: 8px;}
-.content .pl-3 {padding-left: 12px;}
-.content .pl-4 {padding-left: 16px;}
-.content .pl-5 {padding-left: 20px;}
-
-.content .pr-0 {padding-right: 0;}
-.content .pr-1 {padding-right: 4px;}
-.content .pr-2 {padding-right: 8px;}
-.content .pr-3 {padding-right: 12px;}
-.content .pr-4 {padding-right: 16px;}
-.content .pr-5 {padding-right: 20px;}
-
-.field {}
-
-.field label {
-    display: block;
-    padding:4px 8px 4px 0px;
-}
-.field input[type="text"] {
-    padding:6px 6px;
-    outline:none;
-    border: 1px solid #aaa;
-    border-radius: 4px;
-}
-.field input[type="text"], .field input[type="file"], .field select, .field button {
-    height: 24px;
-}
-.field button:not(:last-child) {
-    margin-right: 8px;
-}
-.field canvas {
-    display: block;
-    width: 100%;
-    aspect-ratio: 2 / 1;
-}
-.field canvas.fullscreen {
-    aspect-ratio: none;
-    background-color: #fff;
-    position: absolute;
-    top: 0;
-    left: 0;
-    height: 100vh;
-    width: 100vw;
-    z-index: 100;
-}
-.field canvas:focus {
-    outline: none;
-}
-.field canvas.drag-key-down {
-    cursor: grab;
-}
-.field canvas.drag-key-down.dragging-grid {
-    cursor: grabbing;
-}
-.d-flex {
-    display: flex;
-}
-.d-flex.align-center {
-    align-items: center;
-}
-.spacer {
-    flex-grow: 1;
-}
-
-.raycast-display {
-    position: absolute;
-    display: block;
-    bottom: 0px;
-    left: calc(50% - 400px);
-    z-index: 100;
-    height: 400px;
-    width: 800px;
-    background-color: #000;
-
-    transition: all 0.3s;
-}
-.raycast-display:hover {
-    bottom: 50px;
-    left: calc(50% - 800px);
-    height: 800px;
-    width: 1600px;
-}
-.hide {
-    position: absolute;
-    top: 0;
-    left: 0;
-    pointer-events: none;
-    opacity: 0;
-    height: 1px;
-    width: 1px;
-}
-textarea {
-    display: block;
-    width: 100%;
-    min-height: 200px;
-    box-sizing: border-box;
-}
-textarea:focus {
-    outline: none;
-}
-.legend-input {
-    display: inline-block;
-    padding: 2px 4px;
-    border-radius: 4px;
-    background-color: #aaa;
-    margin-right: 4px;
-}
